Verify tables in MySQL-/MariaDB-database

Checks for tables in a MySQL-/MariaDB-database. Non-existing tables will be compared (case insensitive) to existing tables. When a non-existing table occurs with a prefix ending with an underscore ("_"), the table name will be overwritten with an existing table name listed last in alphabetical order.

certedb_check_tbls(tbls, con = NULL, dbname = "certemmb", info = TRUE)

Arguments

tbls
Tables in the database to be verified.

con
Default is empty, by which the connection is made based on the current user's environment variables: "db.mmb.host", "db.mmb.port", "db.mmb.username" en "db.mmb.password".

dbname
Default is "certemmb". Name of the database to be selected. Will be ignored if con is already an existing connection.

info
Default is TRUE. Display a warning about non-existing tables to be replaced by existing tables.