MMB-gegevens ophalen van MySQL-/MariaDB-database

Gegevens van Certe Medische Microbiologie uit de Certe-database downloaden. De benodigde tabellen worden als LEFT JOIN automatisch toegevoegd op basis van de input bij where en select. Nadien kan met qry de query bekeken worden, die als eigenschap bij het object opgeslagen wordt.

certedb_getmmb haalt orders en resultaten op, met uitslagen.
certedb_getmmb_tat haalt van deze orders de doorlooptijden op (TAT = Turn Around Time).

certedb_getmmb(
  dates = NULL,
  where = NULL,
  select_preset = "mmb",
  preset = "mmb",
  select = NULL,
  add_cols = NULL,
  limit = 1e+07,
  con = NULL,
  dbname = "certemmb",
  info = TRUE,
  first_isolates = FALSE,
  eucast_rules = "all",
  MIC = FALSE,
  zipcodes = FALSE,
  ziplength = 4,
  inhabitants = FALSE,
  tat_hours = FALSE,
  only_real_patients = TRUE,
  only_conducted_tests = TRUE,
  only_validated = FALSE,
  only_show_query = FALSE,
  review_where = as.logical(min(c(getOption("review_where", TRUE), info,
    interactive()))),
  unselect_all_same = FALSE,
  unselect_all_empty = FALSE,
  ...
)

certedb_getmmb_tat(
  dates = NULL,
  where = NULL,
  add_cols = NULL,
  limit = 1e+07,
  con = NULL,
  dbname = "certemmb",
  info = TRUE,
  only_real_patients = TRUE,
  only_conducted_tests = TRUE,
  only_validated = TRUE,
  only_show_query = FALSE,
  ...
)

Arguments

dates
Standaard is dit hele jaar. Geldige opties zijn: leeg (selecteert dit hele jaar) of een vector met 2 datums (bestaande uit oudste en nieuwste datum) of een enkele datum (einddatum wordt laatste dag van dat jaar).

where
Standaard is leeg. Syntax om toe te voegen aan de WHERE-clausule.

R-syntax wordt omgezet naar SQL-syntax, zie Examples. Deze syntax wordt geëvalueerd, dus gebruik van variabelen is ook mogelijk, zoals certedb_getmmb(where = where(o.jaar == mijnjaar)).

NB. Wanneer tabelvelden in meerdere brontabellen voorkomen, moet de tabelreferentie opgegeven worden. Zie Details voor de tabelreferenties en Examples voor voorbeelden.

select_preset, preset
Standaard is "mmb" of "tat" bij doorlooptijden. Variabelen om te selecteren volgens de voorgedefinieerde lijst met variabelen, zie preset.list. Kan ook een vector met meerdere presets zijn. Voor het selecteren van de eerste preset uit de huidige map (bestandsnaam moet beginnen met "preset"), gebruik select_preset = preset.thisfolder().

select
Standaard is leeg. Variabelen om handmatig te selecteren. Deze kunnen het best geselecteerd worden met db en dit overschrijft select_preset.

add_cols
Standaard is leeg. Variabelen in een vector of list om extra te selecteren. Deze kunnen het best geselecteerd worden met db. Gebruik add_cols = c("naam" = db$t.kolom) voor "t.kolom AS naam".

limit
Standaard is 10.000.000. Het aantal rijen dat maximaal opgehaald moet worden.

con
Standaard is leeg, waarmee de verbinding gemaakt wordt op basis van de omgevingsvariabelen van de huidige gebruiker: "DB_HOST", "DB_PORT", "DB_USERNAME" en "DB_PASSWORD".

dbname
Standaard is "certemmb". Naam van de database die geselecteerd moet worden. Wordt genegeerd als con al een bestaande verbinding is.

info
Standaard is TRUE. Printen van voortgang en het uiteindelijke aantal rijen en kolommen dat gedownload is.

first_isolates
Standaard is FALSE. Bepaling van eerste isolaten toevoegen.

eucast_rules
Standaard is "all". EUCAST expert rules toepassen, zie eucast_rules.

MIC
Standaard is FALSE. Toevoegen van MIC's aan alle RSI-kolommen van de de standaard query (i.e. zonder dat select gebruikt wordt).

zipcodes
Standaard is FALSE. Toevoegen van postcodes van de patiënt, als p.postcode. Met name voor gebruik met get_map en plot2.map.

ziplength
Standaard is 4. De gewenste lengte van de postcode.

inhabitants
Standaard is FALSE, maar alleen zinvol als zipcodes = TRUE. Voegt automatisch de data van CBS toe, die opgeslagen liggen in de tabel temporary_certemm_cbs, op basis van de postcodelengte die opgegeven is met ziplength.

tat_hours
Standaard is TRUE. Turn-around-times toevoegen in uren (alleen voor doorlooptijden).

only_real_patients
Standaard is TRUE. Hiermee worden alleen daadwerkelijke patiënten gedownload (geen rondzendingen en testorders). Dit voegt automatisch AND u.is_echte_patient = TRUE toe aan de WHERE.

only_conducted_tests
Standaard is TRUE. Hiermee worden alleen verrichte testen gedownload (geen testen die niet verricht zijn). Dit voegt automatisch AND u.is_verricht = TRUE toe aan de WHERE.

only_validated
Standaard is FALSE bij certedb_getmmb() en TRUE bij certedb_getmmb_tat(). Hiermee worden alleen geautoriseerde uitslagen gedownload. Dit voegt automatisch AND u.uitslag_int <> '(in behandeling)' toe aan de WHERE.

only_show_query
Standaard is FALSE. Draait de query niet, maar toont hem alleen.

review_where
Standaard is FALSE als info = FALSE of de sessie in niet-interactief (RMarkdown). Bij een interactieve sessie (d.w.z. niet in RMarkdown) wordt de WHERE weergegeven die de gebruiker moet accorderen. Dit kan standaard op FALSE gezet worden met getOption("review_where").

unselect_all_same
Standaard is FALSE. Kolommen excluderen die maar 1 unieke waarde bevatten met select(!all_same()).

unselect_all_empty
Standaard is FALSE. Kolommen excluderen die helemaal leeg zijn met select(!all_empty()).

...
Overige parameters die doorgegeven worden aan certedb_query, zoals auto_transform en timezone.

Details

Voor gebruik van where staan hieronder de tabelreferenties. Deze zijn ook beschikbaar via de list db:

a
temporary_certemm_aanvragers_praktijken

aanvr
temporary_certemm_aanvragers_praktijken

beh
temporary_certemm_aanvragers_praktijken

b
temporary_certemm_bacterienlijst

cbs
temporary_certemm_cbs (joint aan certemm_pat.postcode)

d
temporary_certemm_aanmaakdatums

dlt (alleen voor doorlooptijden)
temporary_certemm_doorlooptijden

i
temporary_certemm_isolaten_rsi

i_mic
temporary_certemm_isolaten_mic

l_instelling
temporary_certemm_locaties

l_ontvangst
temporary_certemm_locaties

l_uitvoer
temporary_certemm_locaties

m
temporary_certemm_materiaalgroepen

o
temporary_certemm_orders

p
certemm_pat

pat
temporary_certemm_patienten

t
temporary_certemm_testgroepen

u
temporary_certemm_uitslagen

Locatie wordt dus op drie manieren gekoppeld; als l_instelling (op o.instelling, zoals huisarts), als l_ontvangst (op o.recsitenb, zoals Assen) en als l_uitvoer (op u.uitvafd, zoals Moleculair).

See also

Examples