SQL-query uitvoeren op MySQL-/MariaDB-database

Een SQL-query uitvoeren op een MySQL-/MariaDB-database van bijv. Certe. De output krijgt een qry-attribuut, dat met qry opgehaald kan worden.

certedb_query(
  query,
  limit = 1e+07,
  con = NULL,
  dbname = "certemmb",
  info = TRUE,
  check_mtrl_test_codes = TRUE,
  binary_as_logical = TRUE,
  auto_append_prefix = TRUE,
  auto_transform = TRUE,
  datenames = "en",
  dateformat = "%Y-%m-%d",
  timeformat = "%H:%M",
  decimal.mark = ".",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA"),
  ...
)

Arguments

query
(Bestand met) SQL-tekst die uitgevoerd moet worden. Tabelnamen in de query hoeven geen "temporary_" of "certemm_" te bevatten en zijn hoofdletterongevoelig. Deze query wordt bij het object opgeslagen als eigenschap query dat bekeken kan worden met de functie qry.

limit
Standaard is 10.000.000. Het aantal rijen dat maximaal opgehaald moet worden.

con
Standaard is leeg, waarmee de verbinding gemaakt wordt op basis van de omgevingsvariabelen van de huidige gebruiker: "DB_HOST", "DB_PORT", "DB_USERNAME" en "DB_PASSWORD".

dbname
Standaard is "certemmb". Naam van de database die geselecteerd moet worden. Wordt genegeerd als con al een bestaande verbinding is.

info
Standaard is TRUE. Printen van voortgang van run/fetch en het uiteindelijke aantal rijen en kolommen dat gedownload is.

check_mtrl_test_codes
Standaard is TRUE. Controleren van materiaal- en testcodes in de tijdelijke tabellen van de database. Hiervoor moet ook info = TRUE zijn.

binary_as_logical
Standaard is TRUE. Kolommen die alleen de waarden 0 en/of 1 bevatten, tranformeren naar logical m.b.v. tbl_binary2logical.

auto_append_prefix
Standaard is TRUE. Wanneer tabellen die voorkomen in de query niet bestaan, wordt gezocht naar tabellen met dezelfde naam die beginnen met "temporary_" of "certemm_". Wanneer zo'n tabel gevonden, wordt die gebruikt.

auto_transform
Standaard is TRUE. Automatisch alle gedownloade kolommen transformeren met tbl_guess_columns.

...
Overige ongebruikte parameters

See also

Examples

if (FALSE) {
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties")
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties", limit = 5)

# Door `auto_append_prefix = TRUE` wordt:
locaties <- certedb_query("SELECT * FROM locaties")
# vertaald naar:
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties")
}