Voorgedefinieerde variabelen selecteren

Dit downloadt bestanden uit de map Sys.getenv("R_REFMAP") die de structuur preset_*.sql hebben. Gebruik een of meerdere van deze (zelf te maken) bestanden als parameter select_preset in de functies certedb_getmmb en certedb_getmmb_tat. Zie Details.

preset.list(fullname = FALSE, recursive = FALSE)

preset.thisfolder.list(fullname = FALSE, recursive = TRUE)

preset.exists(name, returnlocation = FALSE)

preset.read(name)

preset.add(name, vector)

preset.thisfolder(recursive = TRUE)

Arguments

fullname

Standaard is FALSE. Bestandsnamen weergeven met prefix prefix_ en suffix .sql.

recursive

Ook onderliggende mappen doorzoeken.

name

Tekst(vector) om te zoeken in de map Sys.getenv("R_REFMAP").

vector

Tekst(vector) met kolomnamen die opgeslagen moet worden als nieuwe preset. Deze kunnen geselecteerd worden met db.

Details

Gebruik:

preset.list voor het weergeven van alle presets in de map Sys.getenv("R_REFMAP");

preset.add voor het toevoegen van een nieuwe preset;

preset.read voor het uitlezen van een preset;

preset.thisfolder.list voor het weergeven van alle presets in de huidige map(pen);

preset.thisfolder voor het weergeven van de eerst gevonden preset in de huidge map;

preset.exists voor toetsen of een preset bestaat.