Run SQL-query on a MySQL-/MariaDB-database

Run an SQL-query on a MySQL-/MariaDB-database from, for example, Certe. Output gets a qry-attribute, which can be retrieved with qry.

Usage,

certedb_query(
  query,
  limit = 1e+07,
  con = NULL,
  dbname = "certemmb",
  info = TRUE,
  check_mtrl_test_codes = TRUE,
  binary_as_logical = TRUE,
  auto_append_prefix = TRUE,
  auto_transform = TRUE,
  datenames = "en",
  dateformat = "%Y-%m-%d",
  timeformat = "%H:%M",
  decimal.mark = ".",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA"),
  ...
)

Arguments

query

(File containing) SQL-text to be run. Table names in the query don't need to contain "temporary_" or "certemm_" and are not case sensitive. This query will be saved with the object as attribute query which can be viewed with the qry function.

limit

Default is 10.000.000. The maximum amount of rows to be retrieved.

con

Default is empty, by which the connection is made based on the current user's environment variables: "db.mmb.host", "db.mmb.port", "db.mmb.username" en "db.mmb.password".

dbname

Default is "certemmb". Name of the database to be selected. Will be ignored if con is already an existing connection.

info

Default is TRUE. Printing of run/fetch progress en the resulting number of rows and columns downloaded.

check_mtrl_test_codes

Default is TRUE. Check on material- and testcodes in the database's temporary tables. Also requires info = TRUE.

binary_as_logical

Default is TRUE. Transform columns only containing 0 and/or 1 values to logical using tbl_binary2logical.

auto_append_prefix

Default is TRUE. When the selected columns in query are non-existent, eponymous tables starting with "temporary_" or "certemm_". When such a table is found, it will be used instead.

auto_transform

Default is TRUE. Automatically transform all downloaded columns with tbl_guess_columns.

...

not used

See also

Examples

if (FALSE) {
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties")
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties", limit = 5)

# With `auto_append_prefix = TRUE`:
locaties <- certedb_query("SELECT * FROM locaties")
# will be transformed to:
locaties <- certedb_query("SELECT * FROM temporary_certemm_locaties")
}